requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-jwt-extended==4.3.1
flask-pymongo==2.3.0
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
werkzeug==2.0.1
gunicorn==20.1.0
pytest==7.0.0
black==22.3.0
flake8==4.0.1
python-jose==3.3.0
bcrypt==3.2.0
marshmallow==3.14.1
geopy==2.2.0
python-dateutil==2.8.2
pytz==2021.3
APScheduler==3.9.1
logging==0.4.9.6 